pyproject.toml
[tool.poetry]
name = "my-agent"
version = "0.1.0"
description = "AI Agent with OpenAI integration"
authors = ["You"]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.12.0"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
pydantic = "^2.8.2"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"